extract out graph collection to a separate component
